colourpicker: normalized float format added
According to the Löve2D wiki, this format is used from version 11.0
https://love2d.org/wiki/love.graphics
so I decided to add it to the colourpicker plugin.

Usage:
Write `f` in the editor, then
`View > Color Picker Window`
or
`Right Click > Insert Colour`
pick your color, press Select, and the color components will be added
with values whitin the range 0 to 1, comma-separated, as required by the
love.graphics module.